Some corrections for Mycelium Android
The above matrix does **not** apply to Mycelium iPhone. That is a completely different app with different features.

Mycelium Android does support some dozens of shitcoins I don't have handy right now. There is an array in the source code though ...

The organization wasn't founded the date presented here with the name it has now. It changed its name and location a few times.